SageMath version 9.8.beta6, Release Date: 2022-12-21

+++ src/sage/libs/ntl/ntl_lzz_pX.pyx
@@ -155,7 +155,7 @@ cdef class ntl_zz_pX():
             self.c = <ntl_zz_pContext_class>modulus
         elif isinstance(modulus, Integer):
             self.c = <ntl_zz_pContext_class>ntl_zz_pContext(modulus)
-        elif isinstance(modulus, long):
+        elif isinstance(modulus, int):
             self.c = <ntl_zz_pContext_class>ntl_zz_pContext(modulus)
         else:
             try:
